Use common helper for merging
